[MaterialLib]  Moved Capillary models to a subdirectory
