+++ MaterialLib/MPL/Properties/CreateProperties.h
@@ -11,6 +11,10 @@
  */
 #pragma once
 
+#include "CapillaryPressureSaturation/CreateSaturationBrooksCorey.h"
+#include "CapillaryPressureSaturation/CreateSaturationLiakopoulos.h"
+#include "CapillaryPressureSaturation/CreateSaturationVanGenuchten.h"
+
 #include "CreateBishopsPowerLaw.h"
 #include "CreateBishopsSaturationCutoff.h"
 #include "CreateConstant.h"
@@ -25,8 +29,8 @@
 #include "CreateRelPermBrooksCorey.h"
 #include "CreateRelPermLiakopoulos.h"
 #include "CreateRelPermVanGenuchten.h"
-#include "CreateSaturationBrooksCorey.h"
+#include "CapillaryPressureSaturation/CreateSaturationBrooksCorey.h"
 #include "CreateSaturationDependentSwelling.h"
-#include "CreateSaturationLiakopoulos.h"
-#include "CreateSaturationVanGenuchten.h"
+#include "CapillaryPressureSaturation/CreateSaturationLiakopoulos.h"
+#include "CapillaryPressureSaturation/CreateSaturationVanGenuchten.h"
 #include "CreateTransportPorosityFromMassBalance.h"
