+++ MaterialLib/MPL/Properties/Properties.h
@@ -14,6 +14,9 @@
 #include "BishopsPowerLaw.h"
 #include "BishopsSaturationCutoff.h"
 #include "Constant.h"
+#include "CapillaryPressureSaturation/SaturationBrooksCorey.h"
+#include "CapillaryPressureSaturation/SaturationLiakopoulos.h"
+#include "CapillaryPressureSaturation/SaturationVanGenuchten.h"
 #include "CurveProperty.h"
 #include "DupuitPermeability.h"
 #include "ExponentialProperty.h"
@@ -24,8 +27,5 @@
 #include "RelPermBrooksCorey.h"
 #include "RelPermLiakopoulos.h"
 #include "RelPermVanGenuchten.h"
-#include "SaturationBrooksCorey.h"
 #include "SaturationDependentSwelling.h"
-#include "SaturationLiakopoulos.h"
-#include "SaturationVanGenuchten.h"
 #include "TransportPorosityFromMassBalance.h"
